+++ Krowi_AchievementFilter/Options.lua
@@ -30,11 +30,18 @@ local defaults = {
             },
             Faction = {
                 Neutral = true,
-                Alliance = addon.Faction.IsAlliance,
-                Horde = addon.Faction.IsHorde
+                Alliance = false,
+                Horde = false
+            },
+            Covenant = {
+                Neutral = true,
+                Kyrian = false,
+                Venthyr = false,
+                NightFae = false,
+                Necrolord = false
             },
             SortBy = {
-                Criteria = addon.L["F_DEFAULT"],
+                Criteria = addon.L["Default"],
                 ReverseSort = false
             }
         },
@@ -343,9 +350,24 @@ local function CreatePanel()
     LibStub("AceConfigDialog-3.0"):AddToBlizOptions(AF_NAME, nil, nil);
 end
 
+local function SetFilters()
+    -- Always reset faction filter
+    addon.Options.db.Filters.Faction.Neutral = true;
+    addon.Options.db.Filters.Faction.Alliance = addon.Faction.IsAlliance;
+    addon.Options.db.Filters.Faction.Horde = addon.Faction.IsHorde;
+
+    -- Always reset covenant filter
+    for covenant, _ in next, addon.Options.db.Filters.Covenant do
+        addon.Options.db.Filters.Covenant[covenant] = false;
+    end
+    addon.Options.db.Filters.Covenant.Neutral = true;
+    addon.Options.db.Filters.Covenant[addon.Objects.Covenant[addon.Objects.Covenant.GetActiveCovenant()]] = true;
+end
+
 -- Load the options
 function options.Load()
     addon.Options = LibStub("AceDB-3.0"):New("Options", defaults, true);
+    addon.Options.SetFilters = SetFilters;
     addon.Options.db = addon.Options.profile;
 
     addon.GUI.ElvUISkin.Load();