+++ Krowi_Databases/DbManager/DbManager/Covenant.cs
@@ -1,0 +1,11 @@
+﻿namespace DbManager
+{
+    public enum Covenant
+    {
+        NoCovenant = 0,
+        Kyrian = 1,
+        Venthyr = 2,
+        NightFae = 3,
+        Necrolord = 4
+    }
+}