+++ Krowi_Databases/DbManager/DbManager/Faction.cs
@@ -4,6 +4,6 @@
     {
         NoFaction = -1,
         Horde = 0,
-        Alliance = 1 
+        Alliance = 1
     }
 }